+++ shared/components/file-input/file-input.tsx
@@ -1,0 +1,58 @@
+'use client';
+
+import type { ComponentProps } from 'react';
+
+import { FileUploadIcon } from '@/shared/assets/icons';
+import { cn } from '@/shared/lib/cn';
+
+import { useFileInput } from './use-file-input';
+import { Button } from '../button';
+
+export type FileInputProps = Omit<ComponentProps<'input'>, 'type' | 'value' | 'onChange'> & {
+  value?: File | null;
+  onChange?: (file: File | null) => void;
+  label?: string;
+};
+
+export const FileInput = ({
+  className,
+  value = null,
+  onChange,
+  disabled,
+  label = '이미지 업로드하기',
+  ...props
+}: FileInputProps) => {
+  const { inputRef, inputId, openPicker, handleFileChange } = useFileInput({ disabled, onChange });
+
+  return (
+    <div className={cn('flex w-full max-w-[360px] flex-col', className)}>
+      <input
+        id={inputId}
+        ref={inputRef}
+        type="file"
+        disabled={disabled}
+        className="sr-only"
+        tabIndex={-1}
+        onChange={handleFileChange}
+        aria-hidden
+        {...props}
+      />
+
+      {value === null ? (
+        <Button
+          variant="secondary"
+          size="small"
+          onClick={openPicker}
+          className="px-4 py-2.5"
+          leftIcon={<FileUploadIcon className="size-5 shrink-0" aria-hidden />}
+        >
+          {label}
+        </Button>
+      ) : (
+        <Button variant="secondary" size="small" onClick={openPicker} className="px-4 py-2.5">
+          {value.name}
+        </Button>
+      )}
+    </div>
+  );
+};
