+++ shared/components/file-input/index.ts
@@ -1,0 +1,2 @@
+export { FileInput } from './file-input';
+export type { FileInputProps } from './file-input';
